NetApp: take care to remove parent snapshot in child share
When creating a share from snapshot (NetApp volume clone),
the parent snapshot is being copied into the child volume.
This is not expected/not being tracked by manila. So we handle the
soft deletion of that snapshot. It cannot be deleted immediately,
because the clone split needs to finish first.

citing from [NetApp documentation](https://docs.netapp.com/us-en/ontap/volumes/split-flexclone-from-parent-task.html):

"Beginning with ONTAP 9.4, for non-guaranteed volumes on AFF systems,
the split operation for FlexClone volumes shares the physical blocks
and does not copy the data. Therefore, splitting of FlexClone volumes
on AFF systems is faster than the FlexClone splitting operation in
other FAS systems in ONTAP 9.4 and later. The improved FlexClone
splitting operation on AFF systems has the following benefits:
Existing Snapshot copies are not deleted"

ONTAP 9.4 is EOL, so we don't need to consider older systems.
Anyway if the snapshot is not existing, because we are having a
space guaranteed volume or being on a FAS system, the snapshot would
not exist and the soft delete (rename) would silently fail.

+++ manila/share/drivers/netapp/dataontap/client/client_cmode.py
@@ -4261,23 +4261,38 @@ class NetAppCmodeClient(client_base.NetAppBaseClient):
         self.send_request('snapshot-delete', api_args)
 
     @na_utils.trace
-    def soft_delete_snapshot(self, volume_name, snapshot_name):
+    def soft_delete_snapshot(self, volume_name, snapshot_name,
+                             soft_only=False):
         """Deletes a volume snapshot, or renames it if delete fails."""
-        try:
-            self.delete_snapshot(volume_name, snapshot_name)
-        except netapp_api.NaApiError:
-            self.rename_snapshot(volume_name,
-                                 snapshot_name,
-                                 DELETED_PREFIX + snapshot_name)
-            msg = _('Soft-deleted snapshot %(snapshot)s on volume %(volume)s.')
-            msg_args = {'snapshot': snapshot_name, 'volume': volume_name}
-            LOG.info(msg, msg_args)
+        def _soft_delete_snapshot(cli, volume_name, snapshot_name):
+            msg_args = {'snap': snapshot_name, 'vol': volume_name}
+            try:
+                cli.rename_snapshot(volume_name,
+                                    snapshot_name,
+                                    DELETED_PREFIX + snapshot_name)
+                msg = _('Soft-deleted snapshot %(snap)s on volume %(vol)s.')
+                LOG.info(msg, msg_args)
+            except netapp_api.NaApiError as e:
+                if e.code == netapp_api.EAPINOTFOUND:
+                    msg = _('Snapshot %(snap)s on volume %(vol)s not found.')
+                    LOG.debug(msg, msg_args)
+                    return
+                else:
+                    raise
 
             # Snapshots are locked by clone(s), so split the clone(s)
-            snapshot_children = self.get_clone_children_for_snapshot(
+            snapshot_children = cli.get_clone_children_for_snapshot(
                 volume_name, snapshot_name)
             for snapshot_child in snapshot_children:
-                self.volume_clone_split_start(snapshot_child['name'])
+                cli.volume_clone_split_start(snapshot_child['name'])
+
+        if soft_only:
+            _soft_delete_snapshot(self, volume_name, snapshot_name)
+        else:
+            try:
+                self.delete_snapshot(volume_name, snapshot_name)
+            except netapp_api.NaApiError:
+                _soft_delete_snapshot(self, volume_name, snapshot_name)
 
     @na_utils.trace
     def prune_deleted_snapshots(self):
